Moved JsonUpload component to a new folder. Moved API documentation into storybook.
