+++ documentation/Admin/TemplateAPI.stories.mdx
@@ -1,4 +1,10 @@
+import { Meta } from "@storybook/addon-docs/blocks";
+
+<Meta title="Admin/Docs: API Documentation" />
+
 # Developer API Guide
+
+> \*Note: This is work-in-progress
 
 ## Template API
