+++ documentation/Forms/FormViewer.stories.mdx
@@ -17,7 +17,6 @@ The form viewer expects a JSON with the following structure.
     "email": ""
   },
   "form": {
-    "id": 4,
     "version": 1,
     "titleEn": "",
     "titleFr": "",
