+++ pages/admin/upload.tsx
@@ -1,4 +1,4 @@
-import JSONUpload from "../../components/forms/JsonUpload/JsonUpload";
+import JSONUpload from "../../components/admin/JsonUpload/JsonUpload";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { requireAuthentication } from "../../lib/auth";
 import React from "react";
